Refactored parser so it owns symbols, typetags, etc
A single parser object contains all the information for a final
module because if there is an error in _any_ of the source code
loaded in a single module, it means the entire module is broken.
Maybe in the future we can add functionality to load modules and
handle errors but that would be more expensive as it would require
setting up a separate parser object, but I guess that's fine
and easy with the way this refactored version works.